Fix VLMs upcasting (#350)
* fix gemma upcasting

* fix internvl

* fix kimi_vl

* fix tests

+++ mlx_vlm/models/multi_modality/vision.py
@@ -208,7 +208,7 @@ class FastGELUActivation(nn.Module):
             0.5
             * input
             * (1.0 + mx.tanh(np.sqrt(2 / np.pi) * (input + 0.044715 * (input**3))))
-        )
+        ).astype(input.dtype)
 
 
 class MLP(nn.Module):
